Seperate provider implementation from EmailNotificationService
- Currently implemented provider: AWSEmailProvider..
- The tests checking AWS SES functionality in NotificationManagerTest were extracted to a new test class, AWSEmailProviderTest.
- Moved Notification constants from interface to dedicated constants file
- Added missing test assertions for Notifications in ZMS
