+++ servers/zms/src/main/java/com/yahoo/athenz/zms/notification/NotificationManager.java
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com.yahoo.athenz.zms;
+package com.yahoo.athenz.zms.notification;
 
 import com.google.common.base.Splitter;
 import com.yahoo.athenz.auth.util.AthenzUtils;
 import com.yahoo.athenz.common.server.notification.Notification;
 import com.yahoo.athenz.common.server.notification.NotificationService;
 import com.yahoo.athenz.common.server.notification.NotificationServiceFactory;
+import com.yahoo.athenz.zms.*;
 import com.yahoo.athenz.zms.store.AthenzDomain;
 import com.yahoo.athenz.zms.utils.ZMSUtils;
 import org.slf4j.Logger;
@@ -31,6 +32,8 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
+
+import static com.yahoo.athenz.common.server.notification.NotificationServiceConstants.*;
 
 public class NotificationManager {
 
@@ -43,7 +46,7 @@ public class NotificationManager {
     private int pendingRoleMemberLifespan;
     private String monitorIdentity;
 
-    NotificationManager(final DBService dbService, final String userDomainPrefix) {
+    public NotificationManager(final DBService dbService, final String userDomainPrefix) {
         this.dbService = dbService;
         this.userDomainPrefix = userDomainPrefix;
         String notificationServiceFactoryClass = System.getProperty(ZMSConsts.ZMS_PROP_NOTIFICATION_SERVICE_FACTORY_CLASS);
@@ -82,8 +85,8 @@ public class NotificationManager {
         }
     }
 
-    void generateAndSendPostPutMembershipNotification(final String domain, final String org,
-             final Role role, Map<String, String> details) {
+    public void generateAndSendPostPutMembershipNotification(final String domain, final String org,
+                                                             final Role role, Map<String, String> details) {
 
         if (!isNotificationFeatureAvailable()) {
             return;
@@ -130,7 +133,7 @@ public class NotificationManager {
 
         // create and process our notification
 
-        Notification notification = createNotification(NotificationService.NOTIFICATION_TYPE_MEMBERSHIP_APPROVAL,
+        Notification notification = createNotification(NOTIFICATION_TYPE_MEMBERSHIP_APPROVAL,
                 recipients, details);
         if (notification != null) {
             notificationService.notify(notification);
@@ -272,7 +275,7 @@ public class NotificationManager {
 
         void sendPendingMembershipApprovalReminders() {
             Set<String> recipients = dbService.getPendingMembershipApproverRoles();
-            Notification notification = createNotification(NotificationService.NOTIFICATION_TYPE_MEMBERSHIP_APPROVAL_REMINDER,
+            Notification notification = createNotification(NOTIFICATION_TYPE_MEMBERSHIP_APPROVAL_REMINDER,
                     recipients, null);
             if (notification != null) {
                 notificationService.notify(notification);
@@ -308,7 +311,7 @@ public class NotificationManager {
                 // notification agent for processing
 
                 Map<String, String> details = processRoleExpiryReminder(domainAdminMap, roleMember);
-                Notification notification = createNotification(NotificationService.NOTIFICATION_TYPE_PRINCIPAL_EXPIRY_REMINDER,
+                Notification notification = createNotification(NOTIFICATION_TYPE_PRINCIPAL_EXPIRY_REMINDER,
                         roleMember.getMemberName(), details);
                 if (notification != null) {
                     notificationService.notify(notification);
@@ -321,7 +324,7 @@ public class NotificationManager {
             for (Map.Entry<String, List<MemberRole>> domainAdmin : domainAdminMap.entrySet()) {
 
                 Map<String, String> details = processMemberExpiryReminder(domainAdmin.getKey(), domainAdmin.getValue());
-                Notification notification = createNotification(NotificationService.NOTIFICATION_TYPE_DOMAIN_MEMBER_EXPIRY_REMINDER,
+                Notification notification = createNotification(NOTIFICATION_TYPE_DOMAIN_MEMBER_EXPIRY_REMINDER,
                         ZMSUtils.roleResourceName(domainAdmin.getKey(), ZMSConsts.ADMIN_ROLE_NAME), details);
                 if (notification != null) {
                     notificationService.notify(notification);
@@ -368,8 +371,8 @@ public class NotificationManager {
                 }
                 domainRoleMembers.add(memberRole);
             }
-            details.put(NotificationService.NOTIFICATION_DETAILS_EXPIRY_ROLES, expiryRoles.toString());
-            details.put(NotificationService.NOTIFICATION_DETAILS_MEMBER, member.getMemberName());
+            details.put(NOTIFICATION_DETAILS_EXPIRY_ROLES, expiryRoles.toString());
+            details.put(NOTIFICATION_DETAILS_MEMBER, member.getMemberName());
 
             return details;
         }
@@ -400,8 +403,8 @@ public class NotificationManager {
                         .append(memberRole.getRoleName()).append(';')
                         .append(memberRole.getExpiration().toString());
             }
-            details.put(NotificationService.NOTIFICATION_DETAILS_EXPIRY_MEMBERS, expiryMembers.toString());
-            details.put(NotificationService.NOTIFICATION_DETAILS_DOMAIN, domainName);
+            details.put(NOTIFICATION_DETAILS_EXPIRY_MEMBERS, expiryMembers.toString());
+            details.put(NOTIFICATION_DETAILS_DOMAIN, domainName);
             return details;
         }
     }
